+++ polylogue/cli/click_command_registration.py
@@ -18,7 +18,6 @@ from polylogue.cli.commands.resume import resume_command
 from polylogue.cli.commands.run import run_command
 from polylogue.cli.commands.schema import schema_command
 from polylogue.cli.commands.tags import tags_command
-from polylogue.cli.commands.watch import watch_command
 
 ROOT_COMMANDS: tuple[click.Command, ...] = (
     run_command,
@@ -34,7 +33,6 @@ ROOT_COMMANDS: tuple[click.Command, ...] = (
     insights_command,
     tags_command,
     schema_command,
-    watch_command,
     diagnostics_group,
 )